Change files extension from .ts to .js Use environment variables from .env file
