+++ cypress-e2e/pages/user-info.page.js
@@ -1,0 +1,6 @@
+class UserInfoPage {
+    getUserFullName() {
+        return cy.get('#adf-userinfo-identity-name-display');
+    }
+}
+export default UserInfoPage;
