--- cypress-e2e/pages/user-info.page.ts
@@ -1,6 +1,0 @@
-class UserInfoPage {
-    getUserFullName() {
-        return cy.get('[id="adf-userinfo-identity-name-display"]');
-    }
-}
-export default UserInfoPage
