+++ cypress-e2e/plugins/index.js
@@ -16,7 +16,15 @@
  * @type {Cypress.PluginConfig}
  */
 // eslint-disable-next-line no-unused-vars
+
+const path = require('path')
+require('dotenv').config({ path: path.resolve(__dirname, '../../.env') })
+
 module.exports = (on, config) => {
-  // `on` is used to hook into various events Cypress emits
-  // `config` is the resolved Cypress config
+
+    config.env.baseUrl = process.env.URL_HOST_ADF;
+    config.env.hr_user = process.env.HR_USER;
+    config.env.hr_user_password = process.env.HR_USER_PASSWORD;
+    console.log(config);
+    return config;
 }
